Cleanup 1 : CLAUDE.md R1-R6, désyncs commentaire↔code, clé de cache P2
CLAUDE.md : R1-R6 écrites comme règles, contestations (b) frontal (borné
autorisé), (c) bf16 (scopé au residual stream Gemma-3), (d) versionnage
(exception artefacts assumée) appliquées.

3 désyncs restants corrigés : fp16/bf16 par défaut (saev5.py), σ-clip
intra-batch vs intra-document (saev5.py), vals float16 vs float32
(fragment_store.py).

check_docs.py : PLACEHOLDER_RE élargi à §<N-À-COMPLÉTER> et <!-- À COMPLÉTER
-->. 5 occurrences réelles trouvées et résolues (docs/architecture.md,
docs/evaluation_protocol.md x2, report/03, RESULTS_TESTS.md §26) : toutes
pointaient vers le même résultat Latent Terms token-level bloqué par l'OOM
de build_token_training_pool (jamais produit) -- remplacées par une mention
honnête du blocage plutôt qu'une référence à un §N qui n'existe pas.

saev5.py : clé de cache P2 (3 sites) intègre EMB_MODEL/EMB_POOLING (R5).

pytest tests/ -q : 165 passed, 1 failed (test_check_docs_clean, prose de
docs/INTERP_EMBED_COVERAGE.md et docs/PDF_APPENDICES_EXTRACT.md, reste à
faire).

+++ src/storage/fragment_store.py
@@ -7,7 +7,9 @@ Avec L0 ≈ 70 (core JumpReLU) + 32 (extra), densité ≈ 4e-4 → CSR ≈ 250 K
 
 Format (torch.save, .pt) :
   { "token_strings": list[str],
-    "rowptr": int64 [T+1], "cols": int32 [nnz], "vals": float16 [nnz],
+    "rowptr": int64 [T+1], "cols": int32 [nnz], "vals": float32 [nnz],
+    # float32, pas float16 : les activations JumpReLU du core (non bornées,
+    # outliers ~1e5) débordent la plage fp16 (max ~65504).
     "shape": (T, d_total),
     "raw_acts": bf16 [T, d_in] (optionnel, supprimé après entraînement SAEBoostResidualSAE) }
